Initial commit
This application originate from the Agriculture and Agri-Food Canada
(AAFC) website http://www.agr.gc.ca/atlas/agpv, which complies to the
Web Content Accessibility Guideline (WCAG) 2.0   Please note that per
the Standard on Web Accessibility defined by the Treasury Board of
Canada, complex maps are part of the list of exclusion (see details).
Complex maps are defined as map that associates information with one or
more locations that cannot be represented by postal addresses or proper
names such as intersection, landmarks, cities, or regions.  An examples
of a complex map is one that shows the location and intensity of
precipitation detected by RADAR network.

This web mapping application is a copy of an official work that is
published by the Government of Canada. ESRI Canada Ltd. has adopted this
application and made it freely available to any and all Canadian Federal
Departments who require a WCAG GeoSpatial Viewer on which to build
Public-facing Web Applications.